dashboard/requirements.txt
streamlit
pandas
plotly
pymssql
python-dotenv
awswrangler
boto3
botocore[crt]